requirements.txt
torch
torchvision
opencv-python
h5py
tqdm
accelerate
tensorboard
tensorboardX
scikit-learn
pandas
kaggle
matplotlib
jupyter
nvitop
gdown
transformers
timm
python-dotenv
lmdb